Profile stages of main and simulation schedules as well
