+++ libs/client_lib/src/lib.rs
@@ -15,13 +15,8 @@ use bevy::{
         component::ComponentId,
         entity::Entity,
         query::Access,
-        schedule::{
-            ParallelSystemDescriptorCoercion, ShouldRun, StageLabel, State, StateError, SystemStage,
-        },
-        system::{
-            Commands, IntoExclusiveSystem, IntoSystem, Local, Res, ResMut, System, SystemId,
-            SystemParam,
-        },
+        schedule::{ParallelSystemDescriptorCoercion, ShouldRun, State, StateError, SystemStage},
+        system::{Commands, IntoSystem, Local, Res, ResMut, System, SystemId, SystemParam},
         world::World,
     },
     log,
@@ -39,10 +34,11 @@ use mr_shared_lib::{
     game::client_factories::VisibilitySettings,
     messages::{EntityNetId, PlayerNetId},
     net::{ConnectionState, ConnectionStatus, MessageId},
+    util::profile_schedule,
     GameState, GameTime, MuddleSharedPlugin, SimulationTime, COMPONENT_FRAMEBUFFER_LIMIT,
     SIMULATIONS_PER_SECOND,
 };
-use std::{borrow::Cow, cell::RefCell, thread_local};
+use std::borrow::Cow;
 
 mod camera;
 mod components;
@@ -53,8 +49,6 @@ mod ui;
 mod visuals;
 
 const TICKING_SPEED_FACTOR: u16 = 10;
-
-thread_local!(static PUFFIN_SCOPES: RefCell<HashMap<Box<dyn StageLabel>, puffin::ProfilerScope>> = RefCell::new(HashMap::default()));
 
 pub struct MuddleClientPlugin;
 
@@ -120,54 +114,7 @@ impl Plugin for MuddleClientPlugin {
             // Not only Egui for builder mode.
             .add_system_set(ui::builder_ui::builder_system_set());
 
-        let stages = builder
-            .app
-            .schedule
-            .iter_stages()
-            .map(|(stage_label, _)| stage_label.dyn_clone())
-            .collect::<Vec<_>>();
-        for stage_label in stages {
-            let puffin_id: &'static str = Box::leak(format!("{:?}", stage_label).into_boxed_str());
-            let before_stage_label: &'static str =
-                Box::leak(format!("puffin_before {:?}", stage_label).into_boxed_str());
-            let after_stage_label: &'static str =
-                Box::leak(format!("puffin_after {:?}", stage_label).into_boxed_str());
-            let stage_label_to_remove = stage_label.dyn_clone();
-
-            builder.add_stage_before(
-                stage_label.dyn_clone(),
-                before_stage_label,
-                SystemStage::parallel().with_system(
-                    (move |_world: &mut World| {
-                        PUFFIN_SCOPES.with(|scopes| {
-                            let mut scopes = scopes.borrow_mut();
-                            scopes.insert(
-                                stage_label.dyn_clone(),
-                                puffin::ProfilerScope::new(
-                                    puffin_id,
-                                    puffin::current_file_name!(),
-                                    "",
-                                ),
-                            );
-                        });
-                    })
-                    .exclusive_system(),
-                ),
-            );
-            builder.add_stage_after(
-                stage_label_to_remove.dyn_clone(),
-                after_stage_label,
-                SystemStage::parallel().with_system(
-                    (move |_world: &mut World| {
-                        PUFFIN_SCOPES.with(|scopes| {
-                            let mut scopes = scopes.borrow_mut();
-                            scopes.remove(&stage_label_to_remove);
-                        });
-                    })
-                    .exclusive_system(),
-                ),
-            );
-        }
+        profile_schedule(&mut builder.app.schedule);
 
         let world = builder.world_mut();
         world
